Options Menu Finished
All the skeleton work is done. Once music and SFX are included with a mixer we just need to change two options in the Blueprints. The Key Bindings screen is blank but Everything needed for each key mapping is done. Once all Mapping is complete, I will add each key mapping button. All is located in my developer folder.