Main Turret Finished
I have the main Turret done. It will track the player when in the cone of vision. The turret is animated to look form left to right, we can tweak the numbers at any point in the future. I am going to try and implement the other stuff we talked about but I wanted t upload the main stuff for now incase I fall behind.