updates!!!!!!!!!!!!!!!!!!!
Got rid of one of my testing maps. Updated the turret to call event upon seeing player. Stop attacking when the player reaches a certain distance from the turret. Created a credits Level and have an actor that when spawned will make the camera fade and load the credits level.