Enemy AI Changes
Gave some changes to pathing with the enemies, as well as making the orcs fire their ranged attacks when not able to reach the player while chasing. Made the orcs projectile destroy on collision with the environment. And fixed some bugs with the battle manager so now they attack the player in slower intervals when grouped up.