Credits Redo
I took the opening credits, removed them from the title screen and made a different level for that. I made the default level the opening credits so that would load on opening the game once packaged instead of playing every time you go to the main menu. I also had to add Karthik to the Credits. I had him in the wrong section.